Bash and jupyter notebook examples of how to create a MiniLibriMix dataset
